nspawn: check the pid in SIGCHLD handler before terminating the container
From: Alban Crequy <alban@endocode.com>

When a process starts systemd-nspawn with exec*() without fork(),
systemd-nspawn can be the parent process of children processes unknown
to systemd-nspawn. It can then receive the signal SIGCHLD for both the
container leader process and the previously started processes. So it
should distinguish them.

This commit is dedicated to the 80s, which are pretty ok ;)